feat: allow filtering of options in multi select
