+++ frontend/bacmet/app/search/components/multi-select-field/multi-select-field.module.css
@@ -1,0 +1,4 @@
+.multiSelectBody {
+    max-height: 20rem;
+    overflow: auto;
+}
